feat: refactor contentful getpage logic
