+++ src/Dfe.PlanTech.Application/Content/Queries/GetPageQuery.cs
@@ -1,28 +1,17 @@
-using AutoMapper;
-using Dfe.PlanTech.Application.Caching.Interfaces;
-using Dfe.PlanTech.Application.Core;
-using Dfe.PlanTech.Application.Exceptions;
-using Dfe.PlanTech.Application.Persistence.Interfaces;
-using Dfe.PlanTech.Application.Persistence.Models;
 using Dfe.PlanTech.Domain.Content.Models;
 using Dfe.PlanTech.Domain.Content.Queries;
-using Dfe.PlanTech.Infrastructure.Application.Models;
-using Microsoft.Extensions.Logging;
 
 namespace Dfe.PlanTech.Application.Content.Queries;
 
-public class GetPageQuery : ContentRetriever, IGetPageQuery
+public class GetPageQuery : IGetPageQuery
 {
-    private readonly ILogger<GetPageQuery> _logger;
-    private readonly IQuestionnaireCacher _cacher;
     private readonly IGetPageQuery _getPageFromDbQuery;
-    readonly string _getEntityEnvVariable = Environment.GetEnvironmentVariable("CONTENTFUL_GET_ENTITY_INT") ?? "4";
+    private readonly IGetPageQuery _getPageFromContentfulQuery;
 
-    public GetPageQuery(GetPageFromDbQuery getPageFromDbQuery, ILogger<GetPageQuery> logger, IQuestionnaireCacher cacher, IContentRepository repository) : base(repository)
+    public GetPageQuery(GetPageFromContentfulQuery getPageFromContentfulQuery, GetPageFromDbQuery getPageFromDbQuery)
     {
-        _cacher = cacher;
-        _logger = logger;
         _getPageFromDbQuery = getPageFromDbQuery;
+        _getPageFromContentfulQuery = getPageFromContentfulQuery;
     }
 
     /// <summary>
@@ -32,52 +21,8 @@ public class GetPageQuery : ContentRetriever, IGetPageQuery
     /// <returns>Page matching slug</returns>
     public async Task<Page?> GetPageBySlug(string slug, CancellationToken cancellationToken = default)
     {
-        var page = await _getPageFromDbQuery.GetPageBySlug(slug, cancellationToken) ?? await GetPageFromContentful(slug, cancellationToken);
-
-        UpdateSectionTitle(page);
+        var page = await _getPageFromDbQuery.GetPageBySlug(slug, cancellationToken) ?? await _getPageFromContentfulQuery.GetPageBySlug(slug, cancellationToken);
 
         return page;
-    }
-
-    /// <summary>
-    /// Retrieves the page for the given slug from Contentful
-    /// </summary>
-    /// <param name="slug"></param>
-    /// <param name="cancellationToken"></param>
-    /// <returns></returns>
-    /// <exception cref="FormatException"></exception>
-    /// <exception cref="KeyNotFoundException"></exception>
-    /// <exception cref="ContentfulDataUnavailableException"></exception>
-    private async Task<Page> GetPageFromContentful(string slug, CancellationToken cancellationToken)
-    {
-        if (!int.TryParse(_getEntityEnvVariable, out int getEntityValue))
-        {
-            throw new FormatException($"Could not parse CONTENTFUL_GET_ENTITY_INT environment variable to int. Value: {_getEntityEnvVariable}");
-        }
-
-        try
-        {
-            var options = new GetEntitiesOptions(getEntityValue,
-                new[] { new ContentQueryEquals() { Field = "fields.slug", Value = slug } });
-            var pages = await repository.GetEntities<Page>(options, cancellationToken);
-
-            var page = pages.FirstOrDefault() ?? throw new KeyNotFoundException($"Could not find page with slug {slug}");
-
-            return page;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "Error fetching page {slug} from Contentful", slug);
-            throw new ContentfulDataUnavailableException($"Could not retrieve page with slug {slug}", ex);
-        }
-    }
-
-    private void UpdateSectionTitle(Page page)
-    {
-        if (page.DisplayTopicTitle)
-        {
-            var cached = _cacher.Cached!;
-            page.SectionTitle = cached.CurrentSectionTitle;
-        }
     }
 }
