+++ src/Dfe.PlanTech.Application/Persistence/Interfaces/IGetEntitiesOptions.cs
@@ -2,6 +2,21 @@ namespace Dfe.PlanTech.Application.Persistence.Interfaces;
 
 public interface IGetEntitiesOptions
 {
-    IEnumerable<IContentQuery>? Queries { get; init; }
-    int Include { get; init; }
+    /// <summary>
+    /// Filter queries (e.g. where field equals value)
+    /// </summary>
+    public IEnumerable<IContentQuery>? Queries { get; init; }
+
+    /// <summary>
+    /// Depth of references to include. 1 = parent only, 2 = parent and child, etc.
+    /// </summary>
+    public int Include { get; init; }
+
+    /// <summary>
+    /// What fields to return from Contentful
+    /// </summary>
+    /// <remarks>
+    /// If null, return all
+    /// </remarks>
+    public IEnumerable<string>? Select { get; set; }
 }
