+++ src/Dfe.PlanTech.Domain/Users/Exceptions/PageNotFoundException.cs
@@ -1,0 +1,9 @@
+namespace Dfe.PlanTech.Domain.Users.Exceptions;
+
+public class PageNotFoundException : Exception
+{
+    public PageNotFoundException(string message)
+        : base(message)
+    {
+    }
+}
