+++ src/Dfe.PlanTech.Infrastructure.Data/CmsDbContext.cs
@@ -196,9 +196,6 @@ public class CmsDbContext : DbContext, ICmsDbContext
 
     }
 
-    public Task<List<T>> ToListAsync<T>(IQueryable<T> queryable, CancellationToken cancellationToken = default)
-    => queryable.ToListAsync(cancellationToken: cancellationToken);
-
     public Task<PageDbEntity?> GetPageBySlug(string slug, CancellationToken cancellationToken = default)
     => Pages.Include(page => page.BeforeTitleContent)
             .Include(page => page.Content)
@@ -208,4 +205,10 @@ public class CmsDbContext : DbContext, ICmsDbContext
 
     public IQueryable<RichTextContentDbEntity> RichTextContentsByPageSlug(string pageSlug)
         => RichTextContents.FromSql($"SELECT * FROM [Contentful].[SelectAllRichTextContentForPageSlug]({pageSlug})");
+
+    public Task<T?> FirstOrDefaultAsync<T>(IQueryable<T> queryable, CancellationToken cancellationToken = default)
+        => queryable.FirstOrDefaultAsync(cancellationToken);
+
+    public Task<List<T>> ToListAsync<T>(IQueryable<T> queryable, CancellationToken cancellationToken = default)
+        => queryable.ToListAsync(cancellationToken: cancellationToken);
 }
