+++ src/Dfe.PlanTech.Web/Routing/CheckAnswersRouter.cs
@@ -4,6 +4,7 @@ using Dfe.PlanTech.Domain.Content.Queries;
 using Dfe.PlanTech.Domain.Responses.Interfaces;
 using Dfe.PlanTech.Domain.Submissions.Enums;
 using Dfe.PlanTech.Domain.Submissions.Interfaces;
+using Dfe.PlanTech.Domain.Users.Exceptions;
 using Dfe.PlanTech.Domain.Users.Interfaces;
 using Dfe.PlanTech.Web.Controllers;
 using Dfe.PlanTech.Web.Models;
@@ -52,7 +53,8 @@ public class CheckAnswersRouter : ICheckAnswersRouter
 
         if (checkAnswerDto == null || checkAnswerDto.Responses == null) throw new DatabaseException("Could not retrieve the answered question list");
 
-        var checkAnswerPageContent = await _getPageQuery.GetPageBySlug(CheckAnswersController.CheckAnswersPageSlug, CancellationToken.None);
+        var checkAnswerPageContent = await _getPageQuery.GetPageBySlug(CheckAnswersController.CheckAnswersPageSlug, cancellationToken) ??
+                                    throw new PageNotFoundException($"Could not find page for {CheckAnswersController.CheckAnswersPageSlug}");
 
         var model = new CheckAnswersViewModel()
         {
